server/requirements.txt
fastapi
uvicorn
cohere
numpy
datasets
motor
pydantic
PyJWT>=2.0
passlib[bcrypt]
